refactor(formatter_core): group `src/` modules by domain (#25527)
Also removed dead code and comments.
