+++ crates/oxc_formatter_core/src/context/mod.rs
@@ -1,3 +1,6 @@
+pub mod options;
+pub mod simple;
+
 use crate::{CoreFormatOptions, IndentStyle, IndentWidth, LineEnding, LineWidth, PrinterOptions};
 
 /// Language-agnostic formatting context trait.
