+++ crates/oxc_formatter_core/src/context/options.rs
@@ -1,4 +1,4 @@
-use std::{fmt, num::ParseIntError, str::FromStr};
+use std::{fmt, str::FromStr};
 
 // Language-agnostic format options shared by all formatter implementations.
 //
@@ -15,8 +15,6 @@ pub enum IndentStyle {
 }
 
 impl IndentStyle {
-    pub const DEFAULT_SPACES: u8 = 2;
-
     /// Returns `true` if this is an [IndentStyle::Tab].
     pub const fn is_tab(self) -> bool {
         matches!(self, IndentStyle::Tab)
@@ -25,19 +23,6 @@ impl IndentStyle {
     /// Returns `true` if this is an [IndentStyle::Space].
     pub const fn is_space(self) -> bool {
         matches!(self, IndentStyle::Space)
-    }
-}
-
-impl FromStr for IndentStyle {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "tab" => Ok(Self::Tab),
-            "space" => Ok(Self::Space),
-            // TODO: replace this error with a diagnostic
-            _ => Err("Unsupported value for this option"),
-        }
     }
 }
 
@@ -72,19 +57,9 @@ impl LineEnding {
         }
     }
 
-    /// Returns `true` if this is a [LineEnding::Lf].
-    pub const fn is_line_feed(self) -> bool {
-        matches!(self, LineEnding::Lf)
-    }
-
     /// Returns `true` if this is a [LineEnding::Crlf].
     pub const fn is_carriage_return_line_feed(self) -> bool {
         matches!(self, LineEnding::Crlf)
-    }
-
-    /// Returns `true` if this is a [LineEnding::Cr].
-    pub const fn is_carriage_return(self) -> bool {
-        matches!(self, LineEnding::Cr)
     }
 }
 
@@ -128,16 +103,6 @@ impl IndentWidth {
 impl Default for IndentWidth {
     fn default() -> Self {
         Self(2)
-    }
-}
-
-impl FromStr for IndentWidth {
-    type Err = ParseFormatNumberError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let value = u8::from_str(s).map_err(ParseFormatNumberError::ParseError)?;
-        let value = Self::try_from(value).map_err(ParseFormatNumberError::TryFromU8Error)?;
-        Ok(value)
     }
 }
 
@@ -203,51 +168,6 @@ impl fmt::Debug for LineWidth {
     }
 }
 
-/// Error type returned when parsing a [LineWidth] or [IndentWidth] from a string fails
-#[expect(clippy::enum_variant_names)]
-pub enum ParseFormatNumberError {
-    /// The string could not be parsed to a number
-    ParseError(ParseIntError),
-    /// The `u16` value of the string is not a valid [LineWidth]
-    TryFromU16Error(LineWidthFromIntError),
-    /// The `u8 value of the string is not a valid [IndentWidth]
-    TryFromU8Error(IndentWidthFromIntError),
-}
-
-impl From<IndentWidthFromIntError> for ParseFormatNumberError {
-    fn from(value: IndentWidthFromIntError) -> Self {
-        Self::TryFromU8Error(value)
-    }
-}
-
-impl From<LineWidthFromIntError> for ParseFormatNumberError {
-    fn from(value: LineWidthFromIntError) -> Self {
-        Self::TryFromU16Error(value)
-    }
-}
-
-impl From<ParseIntError> for ParseFormatNumberError {
-    fn from(value: ParseIntError) -> Self {
-        Self::ParseError(value)
-    }
-}
-
-impl fmt::Debug for ParseFormatNumberError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        fmt::Display::fmt(self, f)
-    }
-}
-
-impl fmt::Display for ParseFormatNumberError {
-    fn fmt(&self, fmt: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ParseFormatNumberError::ParseError(err) => fmt::Display::fmt(err, fmt),
-            ParseFormatNumberError::TryFromU16Error(err) => fmt::Display::fmt(err, fmt),
-            ParseFormatNumberError::TryFromU8Error(err) => fmt::Display::fmt(err, fmt),
-        }
-    }
-}
-
 impl TryFrom<u16> for LineWidth {
     type Error = LineWidthFromIntError;
 
@@ -260,23 +180,18 @@ impl TryFrom<u16> for LineWidth {
     }
 }
 
-impl FromStr for LineWidth {
-    type Err = ParseFormatNumberError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let value = u16::from_str(s).map_err(ParseFormatNumberError::ParseError)?;
-        let value = Self::try_from(value).map_err(ParseFormatNumberError::TryFromU16Error)?;
-        Ok(value)
-    }
-}
-
-/// Error type returned when converting a u16 to a [LineWidth] fails
+/// Error type returned when converting a u8 to an [IndentWidth] fails
 #[derive(Clone, Copy, Debug)]
 pub struct IndentWidthFromIntError(pub u8);
 
 impl fmt::Display for IndentWidthFromIntError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "The indent width should be between {} and {}", LineWidth::MIN, LineWidth::MAX)
+        writeln!(
+            f,
+            "The indent width should be between {} and {}",
+            IndentWidth::MIN,
+            IndentWidth::MAX
+        )
     }
 }
