+++ crates/oxc_formatter_core/src/context/simple.rs
