+++ crates/oxc_formatter_core/src/session/embedded.rs
