+++ crates/oxc_formatter_core/src/session/mod.rs
@@ -1,5 +1,7 @@
 //! Per-run execution unit shared by a root formatter and its embedded children.
 
+pub mod embedded;
+
 use std::sync::Arc;
 
 use oxc_allocator::Allocator;
