+++ crates/oxc_formatter_core/src/write/extensions.rs
