+++ crates/oxc_formatter_core/src/write/mod.rs
@@ -1,3 +1,11 @@
+pub mod arguments;
+pub mod buffer;
+pub mod builders;
+pub mod extensions;
+pub mod formatter;
+mod macros;
+pub mod state;
+
 use crate::{Arguments, Buffer, Formatter};
 
 /// Formatting trait for types that can create a formatted representation. The `biome_formatter` equivalent
